+++ client/stylesheets/main.scss
@@ -13,7 +13,6 @@
 @import 'containers/login-controller';
 @import 'containers/admin-panel';
 @import 'containers/donate';
-@import 'containers/donation-done-box';
 @import 'lib/image-carousal';
 @import 'lib/message-carousal';
 @import 'lib/modal';
